refactor(entity): change `winner` type to `Long`

+++ src/main/java/org/as1iva/entity/Match.java
@@ -22,5 +22,5 @@ public class Match {
     @Column(name = "second_player")
     private Long secondPlayer;
 
-    private Integer winner;
+    private Long winner;
 }
